pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "macropad"
version = "0.1.0"
description = "Async serial controller CLI for ATmega-based macropad projects."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pyserial>=3.5",
  "pillow>=10.0",
  "PySide6>=6.7",
  "qasync>=0.27",
  "psutil>=5.9",
  "keyboard>=0.13; platform_system == 'Windows'",
  "pycaw>=20240210; platform_system == 'Windows'",
  "winsdk>=1.0.0b10; platform_system == 'Windows'",
]
authors = [{ name = "MacroPad" }]
license = { text = "MIT" }
keywords = ["serial", "pyserial", "atmega", "macropad", "asyncio"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Environment :: Win32 (MS Windows)",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Software Development :: Embedded Systems",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-qt>=4.4"]
build = ["pyinstaller>=6.0"]
ui = ["numpy>=1.26", "pygame-ce>=2.5"]

[project.scripts]
macropad = "macropad.cli:main"
macropad-gradient = "macropad.pygame_continuous_ui:main"

[project.gui-scripts]
macropad-controller = "macropad.gui_app:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
